update configure-vm.sh to lock RHEL minor release
When setting up a system with a subscription, lock it to the current
minor release. This prevents the system from being upgraded beyond
that minor release so we can be sure that we are testing on the
version we install. Add a flag to disable this behavior for systems
where we want to test upgrading past the installed version.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>